requirements.txt
binance
pandas
pymongo
python-dotenv
schedule